+++ continual/delay.py
@@ -3,14 +3,15 @@ from typing import Tuple
 import torch
 from torch import Tensor
 
-from .interface import CoModule, Padded, PaddingMode, TensorPlaceholder
+from .module import CoModule, PaddingMode, TensorPlaceholder
+from .utils import temporary_parameter
 
 State = Tuple[Tensor, int]
 
 __all__ = ["Delay"]
 
 
-class Delay(torch.nn.Module, Padded, CoModule):
+class Delay(torch.nn.Module, CoModule):
     """Continual delay modules
 
     This module only introduces a delay in the continual modes, i.e. on `forward_step` and `forward_steps`.
@@ -54,20 +55,9 @@ class Delay(torch.nn.Module, Padded, CoModule):
             and self.state_buffer is not None
         ):
             return (self.state_buffer, self.state_index)
-        else:
-            return None
 
-    def forward_step(self, input: Tensor, update_state=True) -> Tensor:
-        if self._delay == 0:
-            return input
-
-        output, (state_buffer, state_index) = self._forward_step(
-            input, self.get_state()
-        )
-        if update_state:
-            self.state_buffer = state_buffer
-            self.state_index = state_index
-        return output
+    def set_state(self, state: State):
+        self.state_buffer, self.state_index = state
 
     def _forward_step(self, input: Tensor, prev_state: State) -> Tuple[Tensor, State]:
         if prev_state is None:
@@ -89,32 +79,20 @@ class Delay(torch.nn.Module, Padded, CoModule):
 
         return output, (buffer, new_index)
 
+    def forward_step(self, input: Tensor, update_state=True) -> Tensor:
+        if self._delay == 0:
+            return input
+
+        return CoModule.forward_step(self, input, update_state)
+
     def forward_steps(self, input: Tensor, pad_end=False, update_state=True) -> Tensor:
-        outs = []
-        for t in range(input.shape[2]):
-            o = self.forward_step(input[:, :, t], update_state=update_state)
-            if isinstance(o, Tensor):
-                outs.append(o)
+        if self._delay == 0:
+            return input
 
-        if pad_end:
-            # Empty out delay values, but don't save state for the end-padding
-            (tmp_buffer, tmp_index) = self.get_state()
-            tmp_buffer = tmp_buffer.clone()
-            for t, i in enumerate(
-                [self.make_padding(input[:, :, -1]) for _ in range(self.delay)]
-            ):
-                o, (tmp_buffer, tmp_index) = self._forward_step(
-                    i, (tmp_buffer, tmp_index)
-                )
-                if isinstance(o, Tensor):
-                    outs.append(o)
+        with temporary_parameter(self, "padding", (self.delay,)):
+            output = CoModule.forward_steps(self, input, pad_end, update_state)
 
-        if len(outs) > 0:
-            outs = torch.stack(outs, dim=2)
-        else:
-            outs = torch.tensor([])  # pragma: no cover
-
-        return outs
+        return output
 
     def forward(self, input: Tensor) -> Tensor:
         # No delay during regular forward